Move pure helpers to results/ and errors/ with shims for old paths
Relocate formatPayload, truncation, and the classifier portion of errorMapping
to their permanent new homes under src/results/ and src/errors/. The old
src/common/ modules become thin re-export shims so existing tools and tests
continue to compile without modification.

- src/results/format.ts now contains the formatPayload implementation.
- src/results/truncation.ts now contains the truncation primitives.
- src/results/response.ts gains structuredResult and errorResult alongside
  the ResponseFormatter interface scaffolded in Task 1.
- src/errors/classifyError.ts gains the concrete classifyError function and
  an ErrorClassifierImpl implementing the existing interface.
- src/common/formatPayload.ts, truncation.ts, structuredResult.ts, and
  errorMapping.ts are now one-to-three-line shims that re-export from the
  new locations.
- Helper tests are moved via git mv: tests/common/{truncation,structuredResult,
  errorMapping}.test.ts -> tests/{results,errors}/ with imports updated to
  point at the new homes.
